Add files via upload
ESP32 is the brain of Smart Compost System. It collects temperature, humidity,
soil moisture level, and methane gas in the compost box using sensors and control
actions are performed automatically to maintain optimum condition in the compost. Two DC motor are connected to ESP32 which operate the mixer shaft and
upper shutter. DC motors are connected using an L298n motor driver. A water
pump and servo motor are also connected to ESP32 to add water and brown waste
respectively. Smart compost system sends real-time Temperature, humidity, soil
moisture, and methane gas emission to a web page and can be accessed using a
Smartphone or computer through any web browser. Mixer-shaft, water pump, the
upper shutter can also be operated manually through the web page.
Initially, food waste is collected and stored in a container to kick start composting
process by the conventional method of composting. After 2 weeks it is moved to
smart compost system. A mixer-shaft is provided for mixing the compost which
also increases the oxygen flow in the system. The temperature and soil moisture
of the compost is controlled automatically by adding water, brown waste, opening
and closing the upper shutter. soil moisture is maintained by add water if the soil
is dry. If soil moisture is high brown waste is added and mixed to reduce soil
moisture. Similarly, the upper shutter is opened and closed to maintain an optimum
temperature of 300C to 500C within the compost.
The temperature sensor (DHT11) checks the temperature level of the soil and
opens/closes the upper shutter accordingly. If the temperature is below 320C close
the upper shutter. Normal condition shutter is remained open. When the temperature of the compost is below 320C upper shutter is closed hence to provide enough
oxygen for the compost shutter is opened when the temperature reaches 55C and
close when the temperature is below 400C. The moisture level of the soil is measured by using a moisture sensor and a 12V dc water pump is turned ON if moisture
is below the specified range. If moisture is above a specified range compost pile is
mixed. If Still, value moisture value remains high brown organic waste is added to
compost from a vessel places at the inner top of the compost. Compost materials
are also mixed at regular intervals to enhance the decomposition of materials. A
gas sensor MQ4 is also added to identify the level of methane emission. All sensor
data can be monitored and control-action can be taken through a web page.